refactor(namespace): change namespace of mapping.cpp to mapping
Updated the namespace in mapping.cpp to 'mapping' for better organization and consistency within the codebase.
